+++ gsmarena_conf.py
