--- test_smartphone_data.json
@@ -1,69 +1,0 @@
-{
-    "Samsung E1107 Crest Solar": {
-        "Network": {
-            "Technology": "GSM",
-            "2G bands": "GSM 900 / 1800 ",
-            "GPRS": "Class 10",
-            "EDGE": "No"
-        },
-        "Launch": {
-            "Announced": "2009, June. Released 2009, July",
-            "Status": "Discontinued"
-        },
-        "Body": {
-            "Dimensions": "105.2 x 44.2 x 16.4 mm (4.14 x 1.74 x 0.65 in)",
-            "Weight": "77 g (2.72 oz)",
-            "SIM": "Mini-SIM",
-            "other": [
-                "Flashlight"
-            ]
-        },
-        "Display": {
-            "Type": "CSTN, 65K colors",
-            "Size": "1.52 inches, 7.5 cm2 (~16.0% screen-to-body ratio)",
-            "Resolution": "128 x 128 pixels, 1:1 ratio (~119 ppi density)"
-        },
-        "Memory": {
-            "Card slot": "No",
-            "Phonebook": "Yes, up to 500 entries",
-            "Call records": "30 dialed, 30 received, 30 missed calls"
-        },
-        "Camera": {
-            "other": [
-                "No"
-            ]
-        },
-        "Sound": {
-            "Loudspeaker ": "Yes",
-            "3.5mm jack ": "No"
-        },
-        "Comms": {
-            "WLAN": "No",
-            "Bluetooth": "No",
-            "GPS": "No",
-            "Radio": "FM radio",
-            "USB": ""
-        },
-        "Features": {
-            "Sensors": "",
-            "Messaging": "SMS",
-            "Browser": "WAP 2.0/xHTML",
-            "Games": "Yes",
-            "Java": "Yes, MIDP 2.0",
-            "other": [
-                "Organizer\r\nVoice memo\r\nMobile Tracker\r\nPredictive text input"
-            ]
-        },
-        "Battery": {
-            "Type": "Removable Li-Ion 800 mAh battery",
-            "Charging": "Solar panel for battery charging - optional\n",
-            "Stand-by": "Upto 30 min",
-            "Talk time": "Up to 1 min 40 sec"
-        },
-        "Misc": {
-            "Colors": "Black",
-            "SAR EU": "0.56 W/kg (head)     ",
-            "Price": "About 50 EUR"
-        }
-    }
-}
